requirements.txt
pydicom
wandb
torchsummary
torchvision
torchmetrics 
lightning==2.2.3
pillow
matplotlib
pandas
numpy